Extract MethodAndMemberVisitor
