+++ RefactorMCP.ConsoleApp/Tools/MoveMultipleMethods.Tool.cs
@@ -11,55 +11,7 @@ using System.Linq;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Text;
-
-public class MethodAndMemberVisitor : CSharpSyntaxWalker
-{
-    public class MethodInfo
-    {
-        public bool IsStatic { get; set; }
-    }
-
-    public class MemberInfo
-    {
-        public string Type { get; set; } = string.Empty; // "field" or "property"
-    }
-
-    public Dictionary<string, MethodInfo> Methods { get; } = new();
-    public Dictionary<string, MemberInfo> Members { get; } = new();
-
-    public override void VisitMethodDeclaration(MethodDeclarationSyntax node)
-    {
-        var methodName = node.Identifier.ValueText;
-        if (!Methods.ContainsKey(methodName))
-        {
-            Methods[methodName] = new MethodInfo
-            {
-                IsStatic = node.Modifiers.Any(SyntaxKind.StaticKeyword)
-            };
-        }
-    }
-
-    public override void VisitFieldDeclaration(FieldDeclarationSyntax node)
-    {
-        foreach (var variable in node.Declaration.Variables)
-        {
-            var fieldName = variable.Identifier.ValueText;
-            if (!Members.ContainsKey(fieldName))
-            {
-                Members[fieldName] = new MemberInfo { Type = "field" };
-            }
-        }
-    }
-
-    public override void VisitPropertyDeclaration(PropertyDeclarationSyntax node)
-    {
-        var propertyName = node.Identifier.ValueText;
-        if (!Members.ContainsKey(propertyName))
-        {
-            Members[propertyName] = new MemberInfo { Type = "property" };
-        }
-    }
-}
+using RefactorMCP.ConsoleApp.SyntaxRewriters;
 
 [McpServerToolType]
 public static partial class MoveMultipleMethodsTool
